Decrypt the key ourselves since Google won't do it for us.
For some reason attempts to use a secret, encrypted value
via [these instructions](https://cloud.google.com/cloud-build/docs/securing-builds/use-encrypted-secrets-credentials)
won't work. Attempts fail with the opaque error `INVALID_ARGUMENT`
which can apparently be a lot of things due to the disparate number
of causes reported when searching for this on Google.

This PR resolves this issue by committing the encrypted token
as a file and invoking `gcloud kms decrypt` ourserlves. In some ways
I like it more as it's explicit and easy to replicate via local
commands.