Rescale the PDF to fit in the current viewport width. (#18)
This makes changes that allow the PDF to be rescaled to fit the
viewport width.

At smaller resolutions it becomes unreadable, so there's probably
a minimum we'll eventually set. That said we can fry that fish later --
this just gets the code in place so that it doesn't conflict with
the other changes enroute.