Re #28; A first stab at persisting annotations. (#33)
This adds code for persisting a collection of tokens which intersect
with a user's selection.

To do this I incorporated several context instances which act to store
the information we need to do this. This was necessary because the
data is stored at various different levels in the component hierarchy
right now.

I left the styles for this alone, as that can come in a future PR
and should merge with the work @DeNeutoy is doing on the sidebar
now.

The code is pretty rough at this point. I tried to factor things
to make sense, but I anticipate doing the following before this
merges:

- Names. The nomenclature is getting tricky, and verbose. Please
  don't refrain from providing suggestions.
- The top-level `<PDFPage />` component is getting complicated.
  I think I can simplify this by converting some of the domain
  specific models from `interface`s to `class` instances, so that
  logic related to building the appropriate data structures
  lives elsewhere.
- I'm thinking about making a series of small components that
  are responsible for individual pieces of stateful data so
  that it's all not flattened into a single level. The idea is
  we'd have something like:

  ```
  <WithPDFStore>{pdfStore => (
    <WithAnnotationStore>{ annotationStore => (
        ...
    )}
  }</WithPDFStore>
  ```

I'll probably try some revisions tomorrow and next week. This
feels like something we need to discuss and iterate on a lot
before it lands since it's going to define how data moves around
in the UI moving forward.

Co-authored-by: Mark Neumann <mark.neumann.1992@gmail.com>