Fix #35, #34; Don't unintentionally re-render. (#36)
* Fix #35, #34; Don't unintentionally re-render.

This fixes a few bugs that crept in via
02db3d15e08307f2371b8bec85058c7e02e53794.

The fix relies on the use of `useCallback`. Without `useCallback`
the `useEffect()` hook in `<Page />` was getting called too often.
This caused PDF pages to render as selections were being made and
confirmed. A re-render is both expensive and can trigger the error
we were seeing in #35.

The `useCallback` method produces a memoized callback that doesn't
chagne unless the dependencies of that method change. This resolves
the above mentioned issue.

The fix includes a few unrelated changes for code clarity:

- I removed `forwardRef` (we don't need it anymore, but I forgot
  to remove it when removing the reliance on `ref`s in the prior
  PR).
- I renamed `setError` to `onError`. This is both more canonical
  within the community and ensures it's consistent throughout
  the code.

* Don't wrap something that's already an Error in an Error.

* Don't move something we don't have to.