Re #123; Upgrade several UI dependencies.
This started as an effort to upgrade TypeScript. In doing so I
discovered I needed to upgrade eslint too. In doing so I discovered
that this package isn't using our standard mechanism for that, which
caused some irregularities, so I switched this repository to that
too (and reformatted things).

Lastly, I had to upgrade pdfjs too, because of some sort of
incompatibility with the new version of TypeSCript. In doing so I was
able to remove the `@types/pdfjs-dist` package, as the types are now
included.

Sorry for the churn here. This got a lot bigger than I intended.