Add support for editing the number of replicas.
This adds a new property to the `skiff.json` file that makes it
easy to manage the number of replicas.

I also wrote some docs for how things work.

I imagine we'll need to add GPU support too at some point. @soldni
just send me a note when you're ready to do so, and I can help
with that too.